Step Counter is Completed.

+++ app/src/main/java/com/ilaydaberna/imageprocessingbaseddietapp/model/firebase/UserSteps.kt
@@ -1,9 +1,9 @@
 package com.ilaydaberna.imageprocessingbaseddietapp.model.firebase
 
 data class UserSteps(
-        var previousSteps: Float? = null,
-        var totalSteps: Float? = null,
-        var dailySteps: Float? = null,
+        var previousSteps: Int? = null,
+        var totalSteps: Int? = null,
+        var dailySteps: Int? = null,
         var date: Long = 0,
         var userId: String = ""
 )
